requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=py39h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py39h06a4308_0
cffi=1.14.6=py39h400218f_0
cudatoolkit=11.3.1=h2bc3f7f_2
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py39h06a4308_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
importlib-metadata=4.8.1=py39h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py39h06a4308_1
ipython=7.29.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.0=py39h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py39h06a4308_7
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py39h27cfd23_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py39h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.6=py39h06a4308_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.4.3=py39h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pyzmq=22.3.0=py39h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py39h06a4308_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
terminado=0.9.4=py39h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=py39_cu113
tornado=6.1=py39h27cfd23_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py39h06a4308_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
